Fixed scrolling on filter screen Fixed veteran check box behavior Moved qualifying classes from model to new util pkg
