+++ app/src/main/java/edu/gatech/spacebarz/buzzshelter/util/FirebaseAuthManager.java
@@ -1,4 +1,4 @@
-package edu.gatech.spacebarz.buzzshelter.model;
+package edu.gatech.spacebarz.buzzshelter.util;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
